UNet_pre_annotation/requirements.txt
matplotlib==3.7.2
numpy==1.24.4
opencv_python==4.8.0.76
Pillow==9.4.0
Pillow==10.0.1
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
